Add markdown creation script

+++ deutschland_ticket/src/common.py
@@ -1,19 +1,21 @@
 import sqlite3
 from pathlib import Path
+from typing import Optional
 
 import requests
 from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
 
 
-def city_table_connection(table_name: str) -> sqlite3.Connection:
+def city_table_connection(table_name: Optional[str] = None) -> sqlite3.Connection:
     db_path = Path(".").resolve() / "data" / "cities.sqlite"
     conn = sqlite3.connect(db_path)
 
     # FIXME: Connection should not be responsible for this
     # We should wrap our orchestrators in test so that we can
     # change these with confidence
-    conn.execute(f"DROP TABLE IF EXISTS {table_name}")
+    if table_name:
+        conn.execute(f"DROP TABLE IF EXISTS {table_name}")
 
     return conn
